Refactor: Proper Directory Organization
1. Separate lib, app and include in separate directories
2. Add Testing capability
3. Refactor CMakeLists
