+++ CMakeLists.txt
@@ -1,4 +1,4 @@
-cmake_minimum_required(VERSION 3.13.0)
+cmake_minimum_required(VERSION 3.18.0)
 
 # Define the Project Name and Description
 project(
@@ -8,21 +8,34 @@ project(
     DESCRIPTION "A Compiler Frontend for the CRust Language :)")
 
 
-# Do it at some point
-# enable_testing() 
+if(CMAKE_PROJECT_NAME STREQUAL PROJECT_NAME)
 
-# Set required C++ Standard
-set(CMAKE_CXX_STANDARD 20)
-set(CMAKE_CXX_STANDARD_REQUIRED ON)
+    # Set required C++ Standard
+    set(CMAKE_CXX_STANDARD 17)
+    set(CMAKE_CXX_STANDARD_REQUIRED ON)
 
-# Let's ensure -std=c++xx instead of -std=g++xx
-set(CMAKE_CXX_EXTENSIONS OFF)
+    # Let's ensure -std=c++xx instead of -std=g++xx
+    set(CMAKE_CXX_EXTENSIONS OFF)
+    set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3 -Wall -Wextra")
 
+    # Let's nicely support folders in IDEs
+    set_property(GLOBAL PROPERTY USE_FOLDERS ON)
 
-set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -O3 -Wall -Wextra")
+    # Testing only available if this is the main app
+    # Note this needs to be done in the main CMakeLists
+    # since it calls enable_testing, which must be in the
+    # main CMakeLists.
+    include(CTest)
+endif()
 
-# Let's nicely support folders in IDEs
-set_property(GLOBAL PROPERTY USE_FOLDERS ON)
+# FetchContent added in CMake 3.11, downloads during the configure step
+include(FetchContent)
 
-add_subdirectory(src)
-# add_subdirectory(tests)
+# The compiled library code is here
+add_subdirectory(lib)
+
+# The executable code is here
+add_subdirectory(apps)
+
+# The testing code is here
+add_subdirectory(tests)
