+++ lib/parser.cpp
@@ -1,0 +1,1 @@
+#include <crust/parser.hpp>
