+++ lib/utils/errorlogger.cpp
@@ -1,8 +1,8 @@
-#include "errorlogger.hpp"
+#include <crust/utils/errorlogger.hpp>
 
 #include <iostream>
 
-#include "sourceloc.hpp"
+#include <crust/utils/sourceloc.hpp>
 
 using namespace Crust;
