--- src/parser/parser.cpp
@@ -1,1 +1,0 @@
-#include "parser.hpp"
