+++ tests/lexer_tests.cpp
@@ -1,0 +1,3 @@
+#include <gtest/gtest.h>
+
+#include <crust/lexer.hpp>
